backend/requirements.txt
Django==4.2.4
python-dotenv==1.0.0
psycopg2-binary==2.9.6
django-rest-framework==0.1.0
django-prometheus==2.3.1
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
python-keycloak==3.3.0
django-cors-headers==4.2.0
Pillow==10.0.0
celery~=5.3.4
redis
djangorestframework~=3.14.0
requests~=2.31.0
django-oauth-toolkit
drf-spectacular
drf_spectacular_sidecar
docker